+++ packages/cobuild/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./building-packet";
 export * as builtins from "./builtins";
 export * as factory from "./factory";
-export * as json from "./json";
 export * as recipes from "./recipes";
 export * from "./witness-layout";